SetQMconfigtask: add the ability to parametrize the configuration
The parameters dictionary must be returned by the config file by
get_parameters() . Because of this change, the get_config function
must accept the parameters dictionary as an argument.

The parameters dictionary keys are the name of the parameters. The
values are either a single default value or a tuple containing a
default value and a comment/unit displayed to the right in the task
view.